requirements.txt
discord.py==2.3.2
flask==3.0.0
requests==2.31.0
aiohttp==3.9.1
gunicorn==21.2.0
groq
psycopg2-binary
